ï

+++ stuff/libs/qoi/include/stuff/qoi/encoder.hpp
@@ -4,6 +4,7 @@
 #include <stuff/qoi/color.hpp>
 #include <stuff/qoi/detail/common.hpp>
 
+#include <cmath>
 #include <span>
 
 namespace stf::qoi::detail {
@@ -13,7 +14,7 @@ constexpr auto encode_lossless(It out_data, std::span<const color> in_pixels)
   -> expected::expected<It, std::string_view>;
 
 template<typename It>
-constexpr auto encode_lossy(It out_data, std::span<const color> in_pixels, i8 tolerance = 1)
+constexpr auto encode_lossy(It out_data, std::span<const color> in_pixels, double tolerance = 1)
   -> expected::expected<It, std::string_view>;
 
 }  // namespace stf::qoi::detail
